Drag and drop implemented and MemoryLeakDetector added
Also:
- Included several models
- Added ID to exclude from DEBUG

+++ ModuleCamera.cpp
@@ -204,11 +204,7 @@ void ModuleCamera::IncreaseSpeed()
 
 void ModuleCamera::OnWindowResized(int width, int height)
 {
-	// aspect ratio is needed so it does not 
-	//if (width > 0 && height > 0) {
-		frustum.SetVerticalFovAndAspectRatio(frustum.VerticalFov(), ((float) width) / height);
-	//}
-	
+	frustum.SetVerticalFovAndAspectRatio(frustum.VerticalFov(), ((float) width) / (float) height);
 }
 
 void ModuleCamera::SetHorizontalFov(int fov)
